% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad_data_function.R
\name{ad_data}
\alias{ad_data}
\title{Function for retrieving and formatting data from adjacent stations}
\usage{
ad_data(station, normal_max, normal_min, data_id)
}
\arguments{
\item{station}{adjacent station that your are attempting to retrieve data for}

\item{normal_max}{Max year that the normal period spans}

\item{normal_min}{Max year that the normal period spans}

\item{data_id}{column name of the data that you are filling}
}
\description{
Function for retrieving and formatting data from adjacent stations
}
\examples{
\dontrun{}
}
\keyword{internal}
