% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aswe_getstats_function.R
\name{aswe_get_stats}
\alias{aswe_get_stats}
\title{Internal function for calculating quartiles and percentiles for each day for each site from automated sites; called within lapply function in stats_aswe() function
January 2021, Ashlee Jollymore}
\usage{
aswe_get_stats(
  stations,
  survey_period,
  get_year,
  normal_min,
  normal_max,
  force = FALSE
)
}
\arguments{
\item{stations}{station that you are calculating statistics for}

\item{survey_period}{survey period in \%m-\%d format}

\item{get_year}{water year that you are calculatign statistics for}

\item{normal_min}{date for the min normal year}

\item{normal_max}{date of the max normal year}

\item{force}{Whether to overwrite the cached normal data. Defaults to FALSE}
}
\description{
Internal function for calculating quartiles and percentiles for each day for each site from automated sites; called within lapply function in stats_aswe() function
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
