% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aswe_normalratio_prep.R
\name{aswe_normalratio}
\alias{aswe_normalratio}
\title{Function for choosing the right method of filling data for ASWE sites
This is done as the raw data has a lot of issues - taking the mean by day a way to ensure that the data used to model below is the best quality possible
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore}
\usage{
aswe_normalratio(data_soi, stations_adj, data_id, normal_max, normal_min)
}
\arguments{
\item{data_soi}{raw data from the station of interest (soi)}

\item{stations_adj}{a list of adjacent stations}

\item{data_id}{what the data ID is}

\item{normal_max}{year of the max for normal period}

\item{normal_min}{min year of normal period}
}
\description{
Function for choosing the right method of filling data for ASWE sites
This is done as the raw data has a lot of issues - taking the mean by day a way to ensure that the data used to model below is the best quality possible
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
