% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASWE_normals_1981t2010}
\alias{ASWE_normals_1981t2010}
\title{ASWE Normals 1981-2010}
\format{
A data frame with 71 rows and 20 variables:
\describe{
\item{STATIONID}{station id}
\item{STATIONNAME}{name of station}
\item{JAN_1_SWE}{SWE normal for Jan 1}
\item{JAN_1_CODE}{Code for Jan 1 SWE normal}
\item{FEB_1_SWE}{SWE normal for Feb 1}
\item{FEB_1_CODE}{Code for Feb 1 SWE normal}
\item{MAR_1_SWE}{SWE normal for March 1}
\item{MAR_1_CODE}{Code for March 1 SWE normal}
\item{\code{APR_1_SWE}}{double SWE normal for April 1}
\item{\code{APR_1_CODE}}{character Code for April 1 SWE normal}
\item{MAY_1_SWE}{SWE normal for May 1}
\item{MAY_1_CODE}{Code for MAYil 1 SWE normal}
\item{MAY_15_SWE}{SWE normal for May 15}
\item{MAY_15_CODE}{Code for May 15 SWE normal}
\item{JUN_1_SWE}{SWE normal for June 1}
\item{JUN_1_CODE}{Code for June 1 SWE normal}
\item{JUN_15_SWE}{SWE normal for June 15}
\item{JUN_15_CODE}{Code for June 15 SWE normal}
\item{\code{MAX_PEAK_SWE}}{double MAX peak SWE}
\item{\code{MIN_PEAK_SWE}}{double MIN peak SWE}
}
}
\usage{
ASWE_normals_1981t2010
}
\description{
Normals for ASWE sites from 1981-2010. A dataset containing the previously calculated normals for ASWE sites (1981-2010)
}
\keyword{datasets}
