% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_basinaverage_function.R
\name{basin_averaged_swe_interactive}
\alias{basin_averaged_swe_interactive}
\title{Function for getting the daily averaged SWE for interactive plots.
Also includes historic years to include on plotly graph
Note that this function returns DAILY AVERAAGE Swe values, rather than hourly ones (as does the function above)}
\usage{
basin_averaged_swe_interactive(basin_input, exceptions)
}
\arguments{
\item{basin_input}{basin name}

\item{exceptions}{sites to leave out}
}
\description{
Function for getting the daily averaged SWE for interactive plots.
Also includes historic years to include on plotly graph
Note that this function returns DAILY AVERAAGE Swe values, rather than hourly ones (as does the function above)
}
\examples{
\dontrun{}
}
\keyword{internal}
