% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{basin_name}
\alias{basin_name}
\title{Function for assigning a basin name to a station ID}
\usage{
basin_name(id = "All", basin = "All", exceptions = NULL)
}
\arguments{
\item{id}{id of station you are assigning to a basin}

\item{basin}{name of basin}

\item{exceptions}{stations to skip}
}
\description{
Function for assigning a basin name to a station ID
}
\examples{
\dontrun{}
}
\keyword{basin}
\keyword{name}
\keyword{site}
\keyword{snow}
