% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_normal.R
\name{basin_normal}
\alias{basin_normal}
\title{Script for calculating normal for a basin-averaged SWE; Jan 2021 Ashlee Jollymore, RFC}
\usage{
basin_normal(data, normal_max, normal_min)
}
\arguments{
\item{data}{data from basin averaged SWE.}

\item{normal_max}{Max year that the normal period spans}

\item{normal_min}{Min year that the normal period spans}
}
\description{
Script for calculating normal for a basin-averaged SWE; Jan 2021 Ashlee Jollymore, RFC
}
\examples{
\dontrun{}
}
\keyword{internal}
