% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basinaverage_function.R
\name{basin_sites}
\alias{basin_sites}
\title{Function for calculating basin averaged SWE
Used for plotting the basin averaged SWE - both interactive and the static plots}
\usage{
basin_sites(site_input, ...)
}
\arguments{
\item{site_input}{site that you want to get data for}
}
\description{
Function for calculating basin averaged SWE
Used for plotting the basin averaged SWE - both interactive and the static plots
}
\examples{
\dontrun{}
}
\keyword{internal}
