% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_normal_prep_function.R
\name{calc_man_norm}
\alias{calc_man_norm}
\title{Internal function for actually calculating the manual normal data
April 2022, Ashlee Jollymore}
\usage{
calc_man_norm(station, df_time, normal_max, normal_min)
}
\arguments{
\item{df_time}{manual snow data you are calculating normals for}

\item{normal_max}{date of the max normal year}

\item{normal_min}{date for the min normal year}
}
\description{
Internal function for actually calculating the manual normal data
April 2022, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
