% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aswe_datamassage_function.R
\name{data_massage}
\alias{data_massage}
\title{Function for formatting data for SWE normal calculation
Function to put right columns in data and take daily average if hourly data exists for ASWE data. Data formatting is done prior to any data filling or calculation of normals
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore}
\usage{
data_massage(data)
}
\arguments{
\item{data}{data that you are formatting}
}
\description{
Function for formatting data for SWE normal calculation
Function to put right columns in data and take daily average if hourly data exists for ASWE data. Data formatting is done prior to any data filling or calculation of normals
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
