% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaswe_swetopeak_function.R
\name{deltaSWE_datetopeak}
\alias{deltaSWE_datetopeak}
\title{Calculate the difference between the current day and the day of the max SWE. Question of how much more snow could be added given historic record?
December 2021, Ashlee Jollymore, RFC}
\usage{
deltaSWE_datetopeak(data_plot_1, lastday_data, date_peak)
}
\arguments{
\item{data_plot_1}{data you are calculating the potential change in SWE to peak}

\item{lastday_data}{last day with SWE data}

\item{date_peak}{date of peak SWE}
}
\description{
Calculate the difference between the current day and the day of the max SWE. Question of how much more snow could be added given historic record?
December 2021, Ashlee Jollymore, RFC
}
\examples{
\dontrun{}
}
\keyword{internal}
