% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elevation}
\alias{elevation}
\title{Function for returning elevation by station}
\usage{
elevation()
}
\description{
Function for returning elevation by station
}
\examples{
\dontrun{}
}
\keyword{elevations}
\keyword{snow}
