% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fillNA0}
\alias{fillNA0}
\title{Datafilling function
Function for filling in NA with 0 when there is no snow at the ASWE pillow and Aquarius has cut out the negative noise
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore}
\usage{
fillNA0(data, ...)
}
\arguments{
\item{data}{data from the station you are trying to fill in data for}
}
\description{
Datafilling function
Function for filling in NA with 0 when there is no snow at the ASWE pillow and Aquarius has cut out the negative noise
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
