% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snow_stats_function.R
\name{get_snow_stats}
\alias{get_snow_stats}
\title{Function for calculating statistics for a snow station
Diverts into either the workflow for calculating ASWE or manual snow station
November 2021, Ashlee Jollymore}
\usage{
get_snow_stats(
  station_id,
  survey_period,
  get_year,
  normal_min,
  normal_max,
  force = TRUE
)
}
\arguments{
\item{station_id}{ID of the manual station(s) that you want to calculate statistics for}

\item{survey_period}{Month-Day that you want to calculate statistics for.}

\item{get_year}{Water year that you want to retrieve data for. Defaults to "All"}

\item{normal_min}{Water year that the normal date range begins}

\item{normal_max}{Water year that the normal date range ends}

\item{incorrect_sites}{Sites that appear to have incorrect data for the survey period and year selected. Defaults to NA}

\item{incorrect_data}{If there appears to be incorrect sites, then the user can input what the SWE data should be (if catalog is wrong)}
}
\description{
Function for calculating statistics for a snow station
Diverts into either the workflow for calculating ASWE or manual snow station
November 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{Retrieve}
\keyword{snow}
\keyword{statistics}
