% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_monthly_swe.R
\name{getmonthly_deltaSWE}
\alias{getmonthly_deltaSWE}
\title{Function for getting statistics for the monthly delta SWE using historic data
December 2021, Ashlee Jollymore}
\usage{
getmonthly_deltaSWE(id)
}
\arguments{
\item{id}{station ID that you are calculating box plot statistics for}
}
\description{
Function for getting statistics for the monthly delta SWE using historic data
December 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
