% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_monthly_swe.R
\name{getmonthly_deltaSWE_YTD}
\alias{getmonthly_deltaSWE_YTD}
\title{Function for getting statistics for the monthly delta SWE - what the monthly change in SWE has been in this year to date}
\usage{
getmonthly_deltaSWE_YTD(id)
}
\arguments{
\item{id}{station ID that you are calculating box plot statistics for}
}
\description{
Function for getting statistics for the monthly delta SWE - what the monthly change in SWE has been in this year to date
}
\examples{
\dontrun{}
}
\keyword{internal}
