% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveplots_function.R
\name{getSWE_all}
\alias{getSWE_all}
\title{Function for getting current year data}
\usage{
getSWE_all(data)
}
\description{
Function for getting current year data
}
\examples{
\dontrun{}
}
\keyword{internal}
