% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_interactiveplots.R
\name{getSWE_current}
\alias{getSWE_current}
\title{Function for only getting current year data}
\usage{
getSWE_current(data)
}
\description{
Function for only getting current year data
}
\examples{
\dontrun{}
}
\keyword{internal}
