% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aswe_normal_function.R
\name{int_aswenorm}
\alias{int_aswenorm}
\title{Function for actually calculating normals from ASWE stations
January 2021, Ashlee Jollymore}
\usage{
int_aswenorm(data, normal_max, normal_min, data_id)
}
\arguments{
\item{data}{station that you are calculating statistics for}

\item{normal_max}{date of the max normal year}

\item{normal_min}{date for the min normal year}

\item{data_id}{column name of data you are calculating normals for}
}
\description{
Function for actually calculating normals from ASWE stations
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
