% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_normal_aswetomanual_function.R
\name{manual2aswe}
\alias{manual2aswe}
\title{function for checking whether a station had a manual site and for using the manual site to extend the dataset back to calculate normals
IN DEVELOPMENT}
\usage{
manual2aswe(data)
}
\arguments{
\item{data}{data for a aswe site that you want to backfill}
}
\description{
function for checking whether a station had a manual site and for using the manual site to extend the dataset back to calculate normals
IN DEVELOPMENT
}
\examples{
\dontrun{}
}
\keyword{aswe}
\keyword{data}
\keyword{gapfill}
\keyword{internal}
\keyword{manual}
\keyword{with}
