% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_2aswe_function.R
\name{manual_2aswe}
\alias{manual_2aswe}
\title{Function for calculating normals for sites that were converted from manual to aswe sites
June 2022, Ashlee Jollymore}
\usage{
manual_2aswe(id, normal_max, normal_min)
}
\arguments{
\item{id}{station that you are calculating statistics for}

\item{normal_max}{maximum year of normal range}

\item{normal_min}{minimum year of normal range}
}
\description{
Function for calculating normals for sites that were converted from manual to aswe sites
June 2022, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
