% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_datafill_function.R
\name{manual_datafill}
\alias{manual_datafill}
\title{Function for filling data from missing manual stations by adjacent stations using multiple linear regression model
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore}
\usage{
manual_datafill(data, normal_max, normal_min, survey_periods_20, num_obs, ...)
}
\arguments{
\item{data}{data that you are filling in data within}

\item{normal_max}{Max year of the normal time span}

\item{normal_min}{Min year of the normal time span}

\item{survey_periods_20}{survey periods with less than 20 years of data within the nomral time period defined}

\item{num_obs}{number of observations}
}
\description{
Function for filling data from missing manual stations by adjacent stations using multiple linear regression model
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
