% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_get_stats_function.R
\name{manual_get_stats}
\alias{manual_get_stats}
\title{Function for calling withing lapply to calculate statistics for each station - manual stations. Internal function
Part of calculating normals workflow for manual data.
January 2021, Ashlee Jollymore}
\usage{
manual_get_stats(
  data,
  stations,
  survey_period,
  get_year,
  normal_min,
  normal_max,
  force = FALSE,
  ...
)
}
\arguments{
\item{data}{data you are calculating statistics for}

\item{stations}{list of station(s) you want to calculate statistics for}

\item{survey_period}{Survey period you want to calculate statistics for}

\item{get_year}{water year you want to calculate statistics for}

\item{normal_min}{minimum year of normal time span}

\item{normal_max}{max year of normal time span}

\item{force}{whether to force recalculation of normals. Defaults to FALSE}
}
\description{
Function for calling withing lapply to calculate statistics for each station - manual stations. Internal function
Part of calculating normals workflow for manual data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
