% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_normal_function.R
\name{manual_normal}
\alias{manual_normal}
\title{Internal function for calculating normals for manual snow stations
Part of calculating normals workflow for manual data.
January 2021, Ashlee Jollymore}
\usage{
manual_normal(data)
}
\arguments{
\item{data}{data you are calculating statistics for}
}
\description{
Internal function for calculating normals for manual snow stations
Part of calculating normals workflow for manual data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
