% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manual_normal_prep_function.R
\name{manual_normal_prep}
\alias{manual_normal_prep}
\title{Internal function for filling in missing data and calculating normals for manual snow survey data
January 2021, Ashlee Jollymore}
\usage{
manual_normal_prep(data, normal_max, normal_min, data_id, force = FALSE)
}
\arguments{
\item{data}{manual snow data you are calculating normals for}

\item{normal_max}{date of the max normal year}

\item{normal_min}{date for the min normal year}

\item{data_id}{column name you are calculating normals for}

\item{force}{Whether to force the recalculation of the normals. Defaults to FALSE}
}
\description{
Internal function for filling in missing data and calculating normals for manual snow survey data
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
