% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manual_normals_1981t2010}
\alias{manual_normals_1981t2010}
\title{Manual normals 1981-2010}
\format{
A data frame with 407 rows and 38 variables:
\describe{
\item{STATIONID}{station id}
\item{STATIONNAME}{name of station}
\item{JAN_1_DEPTH}{Snow depth normal for Jan 1}
\item{JAN_1_SWE}{SWE normal for Jan 1}
\item{JAN_1_DENSITY}{Density normal for Jan 1}
\item{JAN_1_CODE}{Code for Jan 1 SWE normal}
\item{FEB_1_DEPTH}{Snow depth normal for FEB 1}
\item{FEB_1_SWE}{SWE normal for FEB 1}
\item{FEB_1_DENSITY}{Density normal for FEB 1}
\item{FEB_1_CODE}{Code for FEB 1 SWE normal}
\item{MAR_1_DEPTH}{Snow depth normal for MAR 1}
\item{MAR_1_SWE}{SWE normal for MAR 1}
\item{MAR_1_DENSITY}{Density normal for MAR 1}
\item{MAR_1_CODE}{Code for MAR 1 SWE normal}
\item{APR_1_DEPTH}{Snow depth normal for APR 1}
\item{APR_1_SWE}{SWE normal for APR 1}
\item{APR_1_DENSITY}{Density normal for APR 1}
\item{APR_1_CODE}{Code for APR 1 SWE normal}
\item{MAY_1_DEPTH}{Snow depth normal for MAY 1}
\item{MAY_1_SWE}{SWE normal for MAY 1}
\item{MAY_1_DENSITY}{Density normal for MAY 1}
\item{MAY_1_CODE}{Code for MAY 1 SWE normal}
\item{MAY_15_DEPTH}{Snow depth normal for MAY 15}
\item{MAY_15_SWE}{SWE normal for MAY 15}
\item{MAY_15_DENSITY}{Density normal for MAY 15}
\item{MAY_15_CODE}{Code for MAY 15 SWE normal}
\item{JUN_1_DEPTH}{Snow depth normal for June 1}
\item{JUN_1_SWE}{SWE normal for June 1}
\item{JUN_1_DENSITY}{Density normal for JUNE 1}
\item{JUN_1_CODE}{Code for JUNE 1 SWE normal}
\item{JUN_15_DEPTH}{Snow depth normal for June 15}
\item{JUN_15_SWE}{SWE normal for June 15}
\item{JUN_15_DENSITY}{Density normal for JUNE 15}
\item{JUN_15_CODE}{Code for June 1 SWE normal}
\item{\code{MAX_PEAK_DEPTH}}{double max peak depth}
\item{\code{MIN_PEAK_DEPTH}}{double min peak depth}
\item{\code{MAX_PEAK_SWE}}{double max peak SWE}
\item{\code{MIN_PEAK_SWE}}{double min peak SWE}
}
}
\usage{
manual_normals_1981t2010
}
\description{
SWE normals for manual sites from 1981-2010. A dataset containing the previously calculated normals for manual snow sites (1981-2010)
}
\keyword{datasets}
