% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Station Metadata}
\format{
A data frame with 102 rows and 8 variables:
\describe{
\item{ID}{station id}
\item{BASIN}{basin}
\item{ELEVATION}{Elevation}
\item{LATITUDE}{latitude}
\item{LONGITUDE}{longitude}
\item{YEAR}{year}
\item{NESDIS}{nesdis}
\item{OWNER}{owner}
}
}
\usage{
meta
}
\description{
Meta data for snow stations in BC, containing especially who owns the station
}
\keyword{datasets}
