% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveplots_function.R
\name{plot_climate_aswe}
\alias{plot_climate_aswe}
\title{Function for plotting climate data from ASWE sites
Plot the temp, precip and delta SWE as a separate plot}
\usage{
plot_climate_aswe(path, id, save = "No")
}
\arguments{
\item{path}{path that you want to save plots in}

\item{id}{station ID that you want to plot SWE for}

\item{save}{whether to save the plot and stats. Defaults to 'No'}
}
\description{
Function for plotting climate data from ASWE sites
Plot the temp, precip and delta SWE as a separate plot
}
\examples{
\dontrun{}
}
\keyword{aswe}
\keyword{climate}
\keyword{plot}
