% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveplots_function.R
\name{plot_data}
\alias{plot_data}
\title{Function for formatting data and plotting. Internal function for use within plot_interactive_function (for looping over multiple sites)}
\usage{
plot_data(data_plot_1, stn_id, save, path)
}
\arguments{
\item{data_plot_1}{entire dataset}

\item{stn_id}{station ID you are plotting}

\item{save}{whether to save or not}

\item{path}{path that you want to save plots in}
}
\description{
Function for formatting data and plotting. Internal function for use within plot_interactive_function (for looping over multiple sites)
}
\examples{
\dontrun{}
}
\keyword{internal}
