% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveplots_function.R
\name{plot_interactive_aswe}
\alias{plot_interactive_aswe}
\title{Interactive plots - ASWE sites (plot delta SWE and climate plots)}
\usage{
plot_interactive_aswe(path, id, save = "No")
}
\arguments{
\item{path}{path that you want to save plots in}

\item{id}{station ID that you want to plot SWE for}

\item{save}{whether to save the plot and stats. Defaults to 'No', Options are also TRUE}
}
\description{
Interactive plots - ASWE sites (plot delta SWE and climate plots)
}
\examples{
\dontrun{}
}
\keyword{ASWE}
\keyword{SWE}
\keyword{plot}
