% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_interactiveplots_basin.R
\name{plot_interactive_basin}
\alias{plot_interactive_basin}
\title{Function for plotting SWE across an enture basin basin}
\usage{
plot_interactive_basin(basin, exceptions = NA, path_basin, save)
}
\arguments{
\item{basin}{basin name}

\item{exceptions}{sites to leave out}

\item{path_basin}{path to where basin plots should be saved}

\item{save}{yes or no. Whether to save or not}
}
\description{
Function for plotting SWE across an enture basin basin
}
\examples{
\dontrun{}
}
\keyword{SWE}
\keyword{basin}
\keyword{by}
\keyword{interactive}
\keyword{plot}
