% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveplots_function.R
\name{plot_interactive_manual}
\alias{plot_interactive_manual}
\title{Plotting manual sites}
\usage{
plot_interactive_manual(id, path, save = "No")
}
\arguments{
\item{id}{station ID that you want to plot SWE for (manual)}

\item{path}{path that you want to save plots in}

\item{save}{whether to save the plot and stats. Defaults to 'No'. Can also be TRUE}
}
\description{
Plotting manual sites
}
\examples{
\dontrun{}
}
\keyword{SWE}
\keyword{manual}
\keyword{plot}
