% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltamonthlyswe_boxplot_function.R
\name{plot_monthly_deltaswe}
\alias{plot_monthly_deltaswe}
\title{Do boxplot on data from the data saved
Function gets the monthly change in statistics and creates a boxplot of the data that it then saves.
Feb 2020 Ashlee Jollymore}
\usage{
plot_monthly_deltaswe(id, path_save, save = "no")
}
\arguments{
\item{id}{station ID that you are calculating box plot statistics for}

\item{path_save}{Where the plot itself should be saved}

\item{save}{whether to save the plot or not. Defaults to NO, Options are True or False}
}
\description{
Do boxplot on data from the data saved
Function gets the monthly change in statistics and creates a boxplot of the data that it then saves.
Feb 2020 Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{SWE}
\keyword{delta}
\keyword{monthly}
