% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_max_function}
\alias{rank_max_function}
\title{Function for returning the rank of a value - by max
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore}
\usage{
rank_max_function(x, y)
}
\arguments{
\item{x}{data}

\item{y}{column}
}
\description{
Function for returning the rank of a value - by max
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
