% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_min_function}
\alias{rank_min_function}
\title{Functions for returning the rank of a value - by min
January 2021, Ashlee Jollymore}
\usage{
rank_min_function(x, y)
}
\arguments{
\item{x}{data}

\item{y}{column}
}
\description{
Functions for returning the rank of a value - by min
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
