% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aswe_ratio_function.R
\name{ratio_function}
\alias{ratio_function}
\title{Normal Ratio Method
Function for calculating the difference in mean max SWE for normal ratio calculation for filling in data for ASWE snow stations with 10-20 years of data
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore}
\usage{
ratio_function(station, data_station_oi, data_id, normal_max, normal_min)
}
\arguments{
\item{station}{station you are calculating the normal ration fpr}

\item{data_station_oi}{data from the station of interest; i.e., that you are filling in data for}

\item{data_id}{column name of the data that you are filling}

\item{normal_max}{Max year that the normal period spans}

\item{normal_min}{Max year that the normal period spans}
}
\description{
Normal Ratio Method
Function for calculating the difference in mean max SWE for normal ratio calculation for filling in data for ASWE snow stations with 10-20 years of data
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
