% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_df}
\alias{round_df}
\title{Function for rounding all digits within a data frame to specific number of digits}
\usage{
round_df(x, digits)
}
\arguments{
\item{x}{data frame}

\item{digits}{number of digits you want to round data to}
}
\description{
Function for rounding all digits within a data frame to specific number of digits
}
\examples{
\dontrun{}
}
\keyword{internal}
