% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snow_basins}
\alias{snow_basins}
\title{Return the name of all snow basins within BC}
\usage{
snow_basins()
}
\description{
Return the name of all snow basins within BC
}
\examples{
\dontrun{}
}
\keyword{basin}
\keyword{names}
\keyword{snow}
