% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_datafill_function.R
\name{snow_datafill}
\alias{snow_datafill}
\title{Datafilling function
Internal function for taking aswe data that has between 10-20 years of a normal period and extending that dataset using data from ASWE sites that are within 100 km.
Method uses the normal ratio method to compare site of interest to adjacent sites and calculate estimated dataset.
Station of interest (soi) = station you are trying to find data for.
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore}
\usage{
snow_datafill(data_soi, data_id, normal_max, normal_min, ...)
}
\arguments{
\item{data_soi}{data from the station you are trying to fill in data for}

\item{data_id}{column name of the data that you wish to fill in data for prior to calculating statistics}

\item{normal_max}{Max year that the normal period truncates at.}

\item{normal_min}{Min year that the normal period truncates at.}
}
\description{
Datafilling function
Internal function for taking aswe data that has between 10-20 years of a normal period and extending that dataset using data from ASWE sites that are within 100 km.
Method uses the normal ratio method to compare site of interest to adjacent sites and calculate estimated dataset.
Station of interest (soi) = station you are trying to find data for.
Part of calculating normals workflow for ASWE data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
