% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_stats_function.R
\name{snow_stats}
\alias{snow_stats}
\title{Internal function for calculating daily stats for ASWE sites. Inputs include the data you want to calculate the daily stats for. Including quartiles and stats for all data.
Depends on the SWE_normals() function to calculate normals
January 2021, Ashlee Jollymore}
\usage{
snow_stats(data, normal_min, normal_max, data_id, force = FALSE)
}
\arguments{
\item{data}{Data that you are calculating statistics for.}

\item{normal_min}{date for the min normal year}

\item{normal_max}{date of the max normal year}

\item{data_id}{name of the column containing the data you are calculating statistics on}

\item{force}{Whether to overwrite the cached normal data. Defaults to FALSE}
}
\description{
Internal function for calculating daily stats for ASWE sites. Inputs include the data you want to calculate the daily stats for. Including quartiles and stats for all data.
Depends on the SWE_normals() function to calculate normals
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
