% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_stats_manual_function.R
\name{snow_stats_manual}
\alias{snow_stats_manual}
\title{Internal function for calculating daily stats for manual sites. Inputs include the data you want to calculate the daily stats for. Including quartiles and stats for all data.
Depends on the SWE_normals() function to calculate normals
January 2021, Ashlee Jollymore}
\usage{
snow_stats_manual(data, normal_min, normal_max)
}
\arguments{
\item{data}{Data that you are calculating statistics for.}

\item{normal_min}{date for the min normal year}

\item{normal_max}{date of the max normal year}
}
\description{
Internal function for calculating daily stats for manual sites. Inputs include the data you want to calculate the daily stats for. Including quartiles and stats for all data.
Depends on the SWE_normals() function to calculate normals
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
