% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_stats_table_aswe.R
\name{snow_stats_table_aswe}
\alias{snow_stats_table_aswe}
\title{Function for assembling a table of statistics for designated ASWE snow sites. returns a table of daily (ASWE) and periodic (manual) statistics for a selected station
Snow statistics workflow, Ashlee Jollymore RFC}
\usage{
snow_stats_table_aswe(station_id, parameter_id = "SWE", normal_min, normal_max)
}
\arguments{
\item{station_id}{Station ID that you are assembling a statistics table for}

\item{parameter_id}{Parameter ID that you are getting statistics for. Defaults to "SWE", but can be "Snow_Depth", "Precipitation", "Temperature"}

\item{normal_min}{Year that the normal range will start}

\item{normal_max}{Year that the normal range will end}
}
\description{
Function for assembling a table of statistics for designated ASWE snow sites. returns a table of daily (ASWE) and periodic (manual) statistics for a selected station
Snow statistics workflow, Ashlee Jollymore RFC
}
\examples{
\dontrun{}
}
\keyword{ASWE}
\keyword{internal}
\keyword{statistics}
\keyword{table}
