% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_stats_table_manual.R
\name{snow_stats_table_manual}
\alias{snow_stats_table_manual}
\title{Function for assembling a table of statistics for designated manual snow sites. Returns a table of statistics for a selected manual station
Snow statistics workflow, Ashlee Jollymore RFC}
\usage{
snow_stats_table_manual(station_id, normal_min, normal_max, ...)
}
\arguments{
\item{station_id}{Station ID that you are assembling a statistics table for. Can be one or a string of multiple stations}

\item{normal_min}{Year that the normal range will start}

\item{normal_max}{Year that the normal range will end}
}
\description{
Function for assembling a table of statistics for designated manual snow sites. Returns a table of statistics for a selected manual station
Snow statistics workflow, Ashlee Jollymore RFC
}
\examples{
\dontrun{}
}
\keyword{internal}
