% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_aswe_function.R
\name{stats_aswe}
\alias{stats_aswe}
\title{Functions for calculating and retrieving statistics for ASWE snow sites. This is the function that calls on internal functions to calculate and assemble relevant statistics.
Snow statistics workflow, Ashlee Jollymore RFC. Started 23Oct2018; continually revised since then.}
\usage{
stats_aswe(
  station_id,
  parameter_id = "SWE",
  survey_period = "All",
  get_year = "All",
  normal_min,
  normal_max,
  force = FALSE
)
}
\arguments{
\item{station_id}{Station ID that you are assembling a statistics table for. Can be one or a string of multiple stations}

\item{parameter_id}{Parameter that you are calculating statistics for. Defauls to "SWE", but can be "Snow_Depth", "Precipitation", or "Temperature"}

\item{survey_period}{Month and day that you want to calculate data for. Defaults to "All"}

\item{get_year}{Water year that you want to calculate statistics for. Defaults to "All"}

\item{normal_min}{Year that the normal range will start}

\item{normal_max}{Year that the normal range will end}

\item{force}{Whether to overwrite the cached normal data. Defaults to FALSE}
}
\description{
Functions for calculating and retrieving statistics for ASWE snow sites. This is the function that calls on internal functions to calculate and assemble relevant statistics.
Snow statistics workflow, Ashlee Jollymore RFC. Started 23Oct2018; continually revised since then.
}
\examples{
\dontrun{}
}
\keyword{ASWE}
\keyword{internal}
\keyword{statistics}
