% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_MSWE_function.R
\name{stats_MSWE}
\alias{stats_MSWE}
\title{Function for calculating statistics for a certain time period - i.e., April 1 - for manual stations
Part of calculating normals workflow for manual data.
January 2021, Ashlee Jollymore}
\usage{
stats_MSWE(
  station_id,
  survey_period,
  get_year = "All",
  normal_min,
  normal_max,
  incorrect_sites = NA,
  incorrect_data = NA,
  force = FALSE
)
}
\arguments{
\item{station_id}{ID of the manual station(s) that you want to calculate statistics for}

\item{survey_period}{Month-Day that you want to calculate statistcis for.}

\item{get_year}{Water year that you want to retrieve data for. Defaults to "All"}

\item{normal_min}{Water year that the normal date range begins}

\item{normal_max}{Water year that the normal date range ends}

\item{incorrect_sites}{Sites that appear to have incorrect data for the survey period and year selected. Defaults to NA}

\item{incorrect_data}{If there appears to be incorrect sites, then the user can input what the SWE data should be (if catalog is wrong)}

\item{force}{whether to force recalculation of normals. Defaults to FALSE}
}
\description{
Function for calculating statistics for a certain time period - i.e., April 1 - for manual stations
Part of calculating normals workflow for manual data.
January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{internal}
