% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_monthly_swe.R
\name{SWE_diff_month}
\alias{SWE_diff_month}
\title{Calculate the change in SWE by month for all of the years on record
get data for one station}
\usage{
SWE_diff_month(month_select, data)
}
\arguments{
\item{month_select}{month that you are calculating data for}

\item{data}{historic SWE data for a particular station}
}
\description{
Calculate the change in SWE by month for all of the years on record
get data for one station
}
\examples{
\dontrun{}
}
\keyword{internal}
