% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_monthly_swe.R
\name{SWE_diff_month_current}
\alias{SWE_diff_month_current}
\title{Function for calculating the monthly snow accumulation for the current year}
\usage{
SWE_diff_month_current(month_select, data)
}
\arguments{
\item{month_select}{Month that you are calculating the change in SWE for}

\item{data}{current year SWE data}
}
\description{
Function for calculating the monthly snow accumulation for the current year
}
\examples{
\dontrun{}
}
\keyword{internal}
