% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWE_normals_function.R
\name{SWE_normals}
\alias{SWE_normals}
\title{Script for calculating normals for both ASWE and manual stations. This function calls on the correct function to fill data and calculate normals depending on whether the station is a manual or aswe site.
Part of calculating normals workflow for ASWE data. January 2021, Ashlee Jollymore}
\usage{
SWE_normals(data, normal_max, normal_min, force = FALSE)
}
\arguments{
\item{data}{data from manual or ASWE site(s). Can also be a station ID if the function is being called on its own.}

\item{normal_max}{Max year that the normal period spans}

\item{normal_min}{Min year that the normal period spans}

\item{force}{whether you want to update the cache. Defaults to FALSE, or no}

\item{id}{station ID of the site you are calculating normals for}
}
\description{
Script for calculating normals for both ASWE and manual stations. This function calls on the correct function to fill data and calculate normals depending on whether the station is a manual or aswe site.
Part of calculating normals workflow for ASWE data. January 2021, Ashlee Jollymore
}
\examples{
\dontrun{}
}
\keyword{calculate}
\keyword{for}
\keyword{normal}
\keyword{period}
\keyword{snow}
\keyword{statistics}
